Re-namespace EditorScreenMode
