+++ osu.Game/Screens/Edit/Menus/ScreenSelectionTabControl.cs
@@ -9,8 +9,8 @@ using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.UserInterface;
 using osu.Game.Graphics;
 using osu.Game.Graphics.UserInterface;
+using osu.Game.Screens.Edit.Screens;
 using OpenTK;
-using System.ComponentModel;
 
 namespace osu.Game.Screens.Edit.Menus
 {
@@ -69,17 +69,5 @@ namespace osu.Game.Screens.Edit.Menus
                 Bar.ScaleTo(Vector2.One, transition_length, Easing.OutQuint);
             }
         }
-    }
-
-    public enum EditorScreenMode
-    {
-        [Description("compose")]
-        Compose,
-        [Description("design")]
-        Design,
-        [Description("timing")]
-        Timing,
-        [Description("song")]
-        SongSetup
     }
 }
